replaced markup in Views\Home\Index.cshtml
markup creates a table that is data-bound to the properties of the
detail observable in the view model.

The “<!-- ko -->" syntax lets you include a Knockout binding outside of
a DOM element. In this case, the if binding causes this section of
markup to be displayed only when details is non-null.